requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.5
httpx>=0.27.0
pydantic>=2.9.0
python-multipart>=0.0.6
requests>=2.32.2
motor>=3.6.0
pymongo>=4.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4